Change GamePhase.moveLimit from an int to a byte. Makes serialisation easier. Can revisit later if it needs to be larger.

+++ src/GamePhase.h
@@ -19,7 +19,7 @@ class GamePhase
 			bronstein,
 			us
 		} delayType;
-		unsigned int moveLimit;
+		byte moveLimit;
 };
 
 #endif
